create a unified public header
