+++ examples/example_batch_detector.cpp
@@ -5,8 +5,7 @@
 #include <opencv2/opencv.hpp>
 #include <stdtracer>
 
-#include <pose_detector.h>
-
+#include "pose_detector.h"
 #include "utils.hpp"
 
 // Model flags
