+++ examples/example_stream_detector.cpp
@@ -5,7 +5,7 @@
 #include <opencv2/opencv.hpp>
 #include <stdtracer>
 
-#include <stream_detector.h>
+#include "stream_detector.h"
 
 #include "utils.hpp"
