+++ examples/input.cpp
