+++ examples/input.h
