+++ examples/stream_detector.cpp
@@ -1,4 +1,5 @@
-#include <atomic>
+#include "stream_detector.h"
+
 #include <cstdint>
 #include <thread>
 
@@ -9,12 +10,10 @@
 using ttl::tensor;
 
 #include <channel.hpp>
-#include <paf.h>
-#include <stream_detector.h>
-#include <vis.h>
+#include <openpose-plus.h>
 
 #include "input.h"
-#include "uff-runner.h"
+#include "vis.h"
 
 struct default_inputer : stream_detector::inputer_t {
     const std::vector<std::string> filenames;
