+++ examples/stream_detector.h
@@ -2,6 +2,8 @@
 #include <string>
 #include <vector>
 
+#include <opencv2/opencv.hpp>
+
 #include <human.h>
 
 class stream_detector
